fix: fix personal info component for user profile page
